requirements.txt
torch
torchvision
pillow
numpy
opencv-python